Move `_specs_from_cache*` methods to `url_buildcache`
